requirements.txt
tensorflow
numpy
matplotlib
Pillow
customtkinter
